% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkTrafoComplete}
\alias{checkShapeComplete}
\alias{checkTrafoComplete}
\alias{tToX}
\title{Utility functions for wfgEval}
\usage{
checkTrafoComplete(target.index)

checkShapeComplete(target.index, current.shape)

tToX(t, M, A)
}
\description{
Utility functions for wfgEval
}

