% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parseParams}
\alias{parseParams}
\title{parseParams - utility function for WFG
this is only exported for the tests}
\usage{
parseParams(strList, i)
}
\arguments{
\item{strList}{\cr
.}

\item{i}{\cr
.}
}
\value{
nothing
}
\description{
parseParams - utility function for WFG
this is only exported for the tests
}

