% Generated by roxygen2 (4.0.2): do not edit by hand
\name{randomdirection}
\alias{randomdirection}
\alias{randomsearch}
\title{random search (Rastrigin 1963)
- Initialize individuals by uniform distribution in the search-space
- In each iteration try to improve the individual using a random direction and stepsize
- The random direction is chosen following Marsaglia 1972
- The stepsize is chosen by an exponential distribution, preferring small steps}
\usage{
randomdirection(dim)

randomsearch(func, evals = 1000, inDim = 5)
}
\arguments{
\item{func}{\cr
The function to optimize}

\item{evals}{\cr
The number of evaluations allowed. Divide these into sqrt(evals) individuals and generations (heuristic).}

\item{inDim}{\cr
The number of search-space dimensions.}
}
\value{
The final points and their objective-values.
}
\description{
random search (Rastrigin 1963)
- Initialize individuals by uniform distribution in the search-space
- In each iteration try to improve the individual using a random direction and stepsize
- The random direction is chosen following Marsaglia 1972
- The stepsize is chosen by an exponential distribution, preferring small steps
}
\examples{
v2 = c(0.1, 0.2) # example input vectors
}

