% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sConcave}
\alias{sConcave}
\title{sConcave is the concave pareto frontier \cr}
\usage{
sConcave(x)
}
\arguments{
\item{x}{\cr
The vector in objective space without its last entry.}
}
\value{
The modified vector.
}
\description{
sConcave is the concave pareto frontier \cr
}

