% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sConvex}
\alias{sConvex}
\title{sConvex is the convex pareto frontier \cr}
\usage{
sConvex(x)
}
\arguments{
\item{x}{\cr
The vector in objective space without its last entry.}
}
\value{
The modified vector.
}
\description{
sConvex is the convex pareto frontier \cr
}

