% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sDisc}
\alias{sDisc}
\title{sDisc is a shape of the pareto frontier that is not continuous.}
\usage{
sDisc(x, overall = 1, num.regions = 2, location = 1)
}
\arguments{
\item{x}{\cr
The vector in objective space without its last entry.}

\item{overall}{\cr
The overall form of the pareto frontier, if it is >1 then it is more concave, if <1 then more convex.}

\item{num.regions}{\cr
Number of parts of the pareto frontier.}

\item{location}{\cr
Where the discontinuities are. A larger value moves it to larger values of the first objective.}
}
\value{
The modified vector.
}
\description{
sDisc is a shape of the pareto frontier that is not continuous.
}

