% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sLinear}
\alias{sLinear}
\title{sLinear is the linear pareto frontier \cr}
\usage{
sLinear(x)
}
\arguments{
\item{x}{\cr
The vector in objective space without its last entry.}
}
\value{
The modified vector.
}
\description{
sLinear is the linear pareto frontier \cr
}

