% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sMixed}
\alias{sMixed}
\title{sMixed is a shape of the pareto frontier that has some convex and concave parts\cr}
\usage{
sMixed(x, overall = 1, num = 2)
}
\arguments{
\item{x}{\cr
The vector in objective space without its last entry.}

\item{overall}{\cr
The overall form of the pareto frontier, if it is >1 then it is more convex, if <1 then more concave.}

\item{num}{\cr
Number of parts of the pareto frontier.}
}
\value{
The modified vector.
}
\description{
sMixed is a shape of the pareto frontier that has some convex and concave parts\cr
}

