% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sNone}
\alias{sNone}
\title{sNone is a shape that changes nothing. \cr
Used to move the entries cursor along to change later entries.}
\usage{
sNone(x)
}
\arguments{
\item{x}{\cr
The vector in objective space without its last entry.}
}
\value{
The modified vector.
}
\description{
sNone is a shape that changes nothing. \cr
Used to move the entries cursor along to change later entries.
}

