% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tDecept}
\alias{tDecept}
\title{tDecept creates regions in the search space that have a sub-optimal value but larger area. \cr}
\usage{
tDecept(y, opti.loc = 0.35, aperture = 0.001, deceptive.value = 0.05)
}
\arguments{
\item{y}{\cr
The value of the search space entry to which to apply this to.}

\item{opti.loc}{\cr
The location of the true optimum.}

\item{aperture}{\cr
The size of the opening around the true optimum.}

\item{deceptive.value}{\cr
The value of the sub-optimal areas.}
}
\value{
The modified value.
}
\description{
tDecept creates regions in the search space that have a sub-optimal value but larger area. \cr
}

