% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tFlat}
\alias{tFlat}
\title{tFlat creates a region in search space in which all points have the same objective values \cr}
\usage{
tFlat(y, value = 0.8, from = 0.75, to = 0.85)
}
\arguments{
\item{y}{\cr
The value of the search space entry to which to apply this to.}

\item{value}{\cr
The value of the flat region. The adjacent regions interpolate to this value.}

\item{from}{\cr}

\item{to}{\cr
The region of the search space that is flat, same for every dimension.}
}
\value{
The modified value.
}
\description{
tFlat creates a region in search space in which all points have the same objective values \cr
}

