% Generated by roxygen2 (4.0.2): do not edit by hand
\name{timing}
\alias{timing}
\title{timing}
\usage{
timing(expr, maxSeconds = 4)
}
\arguments{
\item{expr}{\cr
The expression to be timed.}

\item{maxSeconds}{\cr
The maximum number of seconds you want to wait for the averaging. There are heuristics such that normally maxSeconds is not reached.}
}
\value{
number of evaluations per second
}
\description{
The timing function is used alongside the WFG package. It estimates the number of evaluations per second.
}
\examples{
# usage:
timing(runif(100*1000))

# comparing the evaluation speed of wfgEval and wfgWrap on a 5->2 test function of 4 transformations and 1 shape.
spec4 = c(tPoly, tDecept, tMulti, tNonsep, sConvex)
f4 = wfgWrap(2, spec4 )
timing(f4(rep(0.1, 5)))
timing(wfgEval(rep(0.1, 5), 2, spec4 ) )
}

