% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tLinear}
\alias{tLinear}
\title{tLinear creates a linear shift of the true optimum. \cr}
\usage{
tLinear(y, zero.loc = 0.35)
}
\arguments{
\item{y}{\cr
The value of the search space entry to which to apply this to.}

\item{zero.loc}{\cr
The location of the true optimum.}
}
\value{
The modified value.
}
\description{
tLinear creates a linear shift of the true optimum. \cr
}

