% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tMulti}
\alias{tMulti}
\title{tMulti creates many local optima. \cr}
\usage{
tMulti(y, num.minima = 30, hill.size = 95, opti.loc = 0.35)
}
\arguments{
\item{y}{\cr
The value of the search space entry to which to apply this to.}

\item{num.minima}{\cr
The number of the local optima.}

\item{hill.size}{\cr
The size of the hills between the local optima.}

\item{opti.loc}{\cr
The location of the true optimum.}
}
\value{
The modified value.
}
\description{
tMulti creates many local optima. \cr
}

