% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tNone}
\alias{tNone}
\title{tNone is a transformation that changes nothing \cr
Used to move the entries cursor along to change later entries.}
\usage{
tNone(y)
}
\arguments{
\item{y}{\cr
The value of the search space entry to which to apply this to.}
}
\value{
The modified value.
}
\description{
tNone is a transformation that changes nothing \cr
Used to move the entries cursor along to change later entries.
}

