% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tNonsep}
\alias{tNonsep}
\title{tNonsep creates a dependence between objectives. \cr}
\usage{
tNonsep(y, degree = NA)
}
\arguments{
\item{y}{\cr
The value of the search space entry to which to apply this to.}

\item{degree}{\cr
Degree of nonseparability.}
}
\value{
The modified value.
}
\description{
tNonsep creates a dependence between objectives. \cr
}

