% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tParam}
\alias{tParam}
\title{tParam is the parameter-dependent transformation \cr}
\usage{
tParam(y, y.prime = NA, factor = 0.98/49.98, starter = 0.02, ender = 50)
}
\arguments{
\item{y}{\cr
The value of the search space entry to which to apply this to.}

\item{y.prime}{\cr
Is set by the system, it does not occur in the parameters the user writes in the specification}

\item{factor}{\cr
Determines the slope of the ramp.}

\item{starter}{\cr}

\item{ender}{\cr
The influence region.}
}
\value{
The modified value.
}
\description{
tParam is the parameter-dependent transformation \cr
}

