% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tPoly}
\alias{tPoly}
\title{tPoly is the polynomial bias transformation \cr}
\usage{
tPoly(y, alpha = 0.02)
}
\arguments{
\item{y}{\cr
The value of the search space entry to which to apply this to.}

\item{alpha}{\cr
alpha>1 biases toward 0, <1 biases toward 1.}
}
\value{
The modified value.
}
\description{
tPoly is the polynomial bias transformation \cr
}

