% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trueFront}
\alias{trueFront}
\title{trueFront}
\usage{
trueFront(name, spec, removeDominated = TRUE)
}
\arguments{
\item{name}{Name under which to store a .png}

\item{spec}{Specification of the test function to plot}
}
\value{
nothing
}
\description{
A plot showing the true pareto front - making use of wfgEval's parameter true.front
}
\examples{
trueFront("truefrontLinear", c(sConvex))
}

