% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tSum}
\alias{tSum}
\title{tSum creates a dependence between different search-space entries. \cr}
\usage{
tSum(y, i, k, M, from = NA, to = NA, weights = NA)
}
\arguments{
\item{y}{\cr
The value of the search space entry to which to apply this to.}

\item{i}{\cr}

\item{k}{\cr}

\item{M}{\cr
Are set by the system, they do not occur in the parameters the user writes in the specification}

\item{from}{\cr}

\item{to}{\cr
The entries which are made dependent on eachother.}

\item{weights}{\cr
Optional for the weighted sum.}
}
\value{
The modified value.
}
\description{
tSum creates a dependence between different search-space entries. \cr
}

