% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wfgShapes}
\alias{sConcave}
\alias{sConvex}
\alias{sDisc}
\alias{sLinear}
\alias{sMixed}
\alias{sNone}
\alias{wfgShapes}
\title{WFG Shapes}
\usage{
wfgShapes()

sNone(x)

sLinear(x)

sConvex(x)

sConcave(x)

sMixed(x, overall = 1, num = 2)

sDisc(x, overall = 1, num = 2, location = 1)
}
\arguments{
\item{x}{\cr
  The vector in objective space without its last entry.}

\item{overall}{\cr
The overall form of the pareto frontier, if it is >1 then it is more convex, if <1 then more concave.}

\item{num}{\cr
Number of parts of the pareto frontier.}

\item{location}{\cr
Where the discontinuities are. A larger value moves it to larger values of the first objective.}

\item{overall}{\cr
The overall form of the pareto frontier, if it is >1 then it is more concave, if <1 then more convex.}
}
\value{
The modified vector.
}
\description{
sNone is a shape that changes nothing. It is used to move the entries cursor along to change later entries.\cr\cr
sLinear is the linear pareto frontier. \cr\cr
sConvex is the convex pareto frontier. \cr\cr
sConcave is the concave pareto frontier. \cr\cr
sMixed is a shape of the pareto frontier that has some convex and concave parts. \cr\cr
sDisc is a shape of the pareto frontier that is not continuous.\cr\cr
}

