% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wfgSuite}
\alias{wfgSuite}
\title{this is a deferred initialization to bring definitions of some of the WFG example problems into the global environment
they are wfgXspec, wfgX, wfgXt where X is 1, 4, 5, 6, 7, or 9
wfgXSpec is only the declarative specification,
wfgX is the actual evaluating function
wfgXt is its true-front version}
\usage{
wfgSuite(n = 5, k = 3, M = 2)
}
\arguments{
\item{n}{\cr
Number of search-space dimensions}

\item{k}{\cr
Number position-related parameters}

\item{M}{\cr
Number of objective-space dimensions}
}
\value{
nothing
}
\description{
this is a deferred initialization to bring definitions of some of the WFG example problems into the global environment
they are wfgXspec, wfgX, wfgXt where X is 1, 4, 5, 6, 7, or 9
wfgXSpec is only the declarative specification,
wfgX is the actual evaluating function
wfgXt is its true-front version
}
\examples{
wfgSuite()
wfg1(rep(0.1, 5))
}

