% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wfgTrafos}
\alias{tDecept}
\alias{tFlat}
\alias{tLinear}
\alias{tMulti}
\alias{tNone}
\alias{tNonsep}
\alias{tParam}
\alias{tPoly}
\alias{tSum}
\alias{wfgTrafos}
\title{WFG Transformations}
\usage{
wfgTrafos()

tNone(y)

tPoly(y, alpha = 0.02)

tFlat(y, value = 0.8, from = 0.75, to = 0.85)

tParam(y, y.prime = NA, factor = 0.98/49.98, starter = 0.02, ender = 50)

tLinear(y, zero.loc = 0.35)

tDecept(y, opti.loc = 0.35, aperture = 0.001, deceptive.value = 0.05)

tMulti(y, num.minima = 30, hill.size = 95, opti.loc = 0.35)

tSum(y, i, k, M, from = NA, to = NA, weights = NA)

tNonsep(y, degree = NA)
}
\arguments{
\item{y}{The value of the search space entry to which to apply this to.
\cr}

\item{alpha}{tPoly: alpha>1 biases toward 0, <1 biases toward 1.
\cr}

\item{value}{tFlat: The value of the flat region. The adjacent regions interpolate to this value.}

\item{from}{}

\item{to}{tFlat: The region of the search space that is flat, same for every dimension.
\cr}

\item{y.prime}{tParam: Is set by the system, it does not occur in the parameters the user writes in the specification}

\item{factor}{tParam: Determines the slope of the ramp.}

\item{starter}{}

\item{ender}{tParam: The influence region.
\cr}

\item{zero.loc}{tLinear: The location of the true optimum.
\cr}

\item{opti.loc}{tDecept: The location of the true optimum.}

\item{aperture}{tDecept: The size of the opening around the true optimum.}

\item{deceptive.value}{tDecept: The value of the sub-optimal areas.
\cr}

\item{opti.loc}{tMulti: The location of the true optimum.}

\item{num.minima}{tMulti: The number of the local optima.}

\item{hill.size}{tMulti: The size of the hills between the local optima.
\cr}

\item{i}{}

\item{k}{}

\item{M}{tSum: Are set by the system, they do not occur in the parameters the user writes in the specification}

\item{from}{}

\item{to}{tSum: The entries which are made dependent on eachother.}

\item{weights}{tSum: Optional for the weighted sum.
\cr}

\item{degree}{tNonsep: Degree of nonseparability.
\cr}
}
\value{
The modified value.
}
\description{
tNone is a transformation that changes nothing. It is used to move the entries cursor along to change later entries.\cr
tPoly is the polynomial bias transformation.\cr
tFlat creates a region in search space in which all points have the same objective values.\cr
tParam is the parameter-dependent transformation.\cr
tLinear creates a linear shift of the true optimum.\cr
tDecept creates regions in the search space that have a sub-optimal value but larger area.\cr
tMulti creates many local optima.\cr
tSum creates a dependence between different search-space entries.\cr
tNonsep creates a dependence between objectives.\cr
}

