% Generated by roxygen2 (4.0.2): do not edit by hand
\name{withDominated}
\alias{baGraphicsWithDominated}
\alias{withDominated}
\title{withDominated is a visualization used alongside the wfg package. Showing the objective space but complemented by the dominated individuals.
baGraphicsWithDominated is a wrapper taking a filename for output and a specification for the test-function.}
\usage{
withDominated(func)

baGraphicsWithDominated(name, spec)
}
\arguments{
\item{func}{\cr
The function to plot}
}
\value{
nothing
}
\description{
withDominated is a visualization used alongside the wfg package. Showing the objective space but complemented by the dominated individuals.
baGraphicsWithDominated is a wrapper taking a filename for output and a specification for the test-function.
}
\examples{
withDominated(wfgWrap(2, c(sConvex)))
baGraphicsWithDominated("withDominatedTFlat", c(tFlat, from=0.4, to=0.9, sLinear))
}

