% Generated by roxygen2 (4.0.2): do not edit by hand
\name{randomdirection}
\alias{randomdirection}
\alias{randomsearch}
\title{random search}
\usage{
randomdirection(dim)

randomsearch(func, evals = 1000, inDim = 5)
}
\arguments{
\item{func}{\cr
The function to optimize}

\item{evals}{\cr
The number of evaluations allowed. Will be divided into sqrt(evals) individuals and generations (heuristic).}

\item{inDim}{\cr
The number of search-space dimensions.}
}
\value{
The final points and their objective-values.
}
\description{
An example algorithm used alongside the WFG package with the goal to be as simple as possible while still resembling better algorithms in that it is probabilistic, population-based and iterative (Rastrigin 1963).\cr\cr
It follows the steps:\cr
- Initialize individuals by uniform distribution in the search-space. \cr
- In each iteration try to improve each individual using a random direction and stepsize. \cr\cr
The random direction is chosen with Marsaglia's method (1972). \cr
The stepsize is chosen by an exponential distribution, preferring small steps. \cr
}

