% Generated by roxygen2 (4.0.2): do not edit by hand
\name{timing}
\alias{timing}
\title{timing function used alongside the wfg package
number of evaluations per second}
\usage{
timing(expr, maxSeconds = 4)
}
\arguments{
\item{expr}{\cr
the expression to be timed.}

\item{maxSeconds}{\cr
the maximum number of seconds you want to wait for the averaging. There are heuristics such that normally maxSeconds is not reached.}
}
\value{
number of evaluations per second
}
\description{
timing function used alongside the wfg package
number of evaluations per second
}
\examples{
# usage:
# timing(runif(100*1000))

# some verification:
rand.size = 100*1000
num.evals = timing(runif(rand.size))
num.evals
p = proc.time()[1]
for (i in 1:num.evals) runif(rand.size)
duration = proc.time()[1] - p # this should approx. 1 sec.
if ( abs(duration - 1) > 0.1 ) stop("timing function inaccurate")

# comparing the evaluation speed of wfgEval and wfgWrap on a 5->2 test function of 4 transformations and 1 shape.
spec4 = c(tPoly, tDecept, tMulti, tNonsep, sConvex)
f4 = wfgWrap(2, spec4 )
timing(f4(rep(0.1, 5)))
timing(wfgEval(rep(0.1, 5), 2, spec4 ) )
}

