% Generated by roxygen2 (4.0.2): do not edit by hand
\name{toPlot}
\alias{toPlot}
\alias{toPlot22}
\title{A visualization used alongside the wfg package.
Showing for each point in the discretized search-space which objective-values will be realized by it as color.
At the moment the first two search-space dimensions are selected (this should be parameterized in the future)
it is possible to show an RGB- or red-blue plot (chosen because of common red-green blindness).
red-blue: red is a large value in the first objective, blue in the second.
Green dots show the non-dominated individuals of a run of random-search, it is suprisingly instructive.
toPlot22 is a 2->2 dimensional wrapper taking a wfg spec instead of an arbitrary function.}
\usage{
toPlot(func, ranked = TRUE, inDim = 5, ...)

toPlot22(spec, ranked = TRUE, ...)
}
\arguments{
\item{func}{\cr
The function to be shown}

\item{ranked}{\cr
Decides whether the colors are chosen based on ranked objective-values. This is recommended to distinguish colors but it does not allow interpretation of their absolute values.}

\item{inDim}{\cr
Number of search-space dimensions; needed for the optimization algorithm (e.g. random search)}
}
\value{
nothing
}
\description{
A visualization used alongside the wfg package.
Showing for each point in the discretized search-space which objective-values will be realized by it as color.
At the moment the first two search-space dimensions are selected (this should be parameterized in the future)
it is possible to show an RGB- or red-blue plot (chosen because of common red-green blindness).
red-blue: red is a large value in the first objective, blue in the second.
Green dots show the non-dominated individuals of a run of random-search, it is suprisingly instructive.
toPlot22 is a 2->2 dimensional wrapper taking a wfg spec instead of an arbitrary function.
}
\examples{
toPlot(wfgWrap(2, c(tDecept, aperture = 0.25, sLinear) ) )
toPlot22( c(tFlat, from=0.4, to=0.9, sLinear) )
toPlot22( c(tMulti, hill.size=0, num.minima=100, sLinear) )
toPlot22( c(tNonsep, degree=10, sLinear) )
}

