% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trueFront}
\alias{trueFront}
\title{A plot showing the true pareto front
making use of wfgEval's parameter true.front}
\usage{
trueFront(name, spec, removeDominated = TRUE)
}
\arguments{
\item{name}{\cr
Name under which to store the png}

\item{spec}{\cr
Specification of the test function to plot}
}
\value{
nothing
}
\description{
A plot showing the true pareto front
making use of wfgEval's parameter true.front
}
\examples{
trueFront("truefrontLinear", c(sConvex))
}

