% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wfgShape}
\alias{wfgShape}
\title{wfgShape - used by wfgEval to apply a shape to part of the objective-space vector}
\usage{
wfgShape(x, current, start.index, apply.length, params)
}
\arguments{
\item{x}{The point at which to evaluate the shape function. It has dimension of search-space.}

\item{current}{The current shape function specified by the user}

\item{start.index}{Specifying to which entries to apply the shape to.}

\item{apply.length}{see start.index}

\item{params}{List of parameters to this shape}
}
\value{
The vector x to which the shape has been applied. Those entries that were changed now have their final objective value, but more shapes can follow for the later indices.
}
\description{
wfgShape - used by wfgEval to apply a shape to part of the objective-space vector
}

