% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wfgSuite}
\alias{wfgSuite}
\title{wfgSuite}
\usage{
wfgSuite(n = 5, k = 3, M = 2)
}
\arguments{
\item{n}{Number of search-space dimensions}

\item{k}{Number position-related parameters}

\item{M}{Number of objective-space dimensions}
}
\value{
nothing
}
\description{
Some of the example problems given by the WFG 2006 paper. \cr
Implementation note: This is a deferred initialization which brings the definitions into the global environment. \cr
They are wfgXspec, wfgX, wfgXt where X is 1, 4, 5, 6, 7, or 9. \cr
wfgXSpec is only the declarative specification,  \cr
wfgX is the actual evaluating function, \cr
wfgXt is its true-front version. \cr
}
\examples{
wfgSuite()
wfg1(rep(0.1, 5))
}

