% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wfgTransformation}
\alias{wfgTransformation}
\title{wfgTransformation - used by wfgEval to apply a transformation to part of the search-space vector}
\usage{
wfgTransformation(M, current, t_, start.index, apply.length, n, k, params)
}
\arguments{
\item{M}{The number of objectives}

\item{current}{The current transformation function specified by the user}

\item{start.index}{}

\item{apply.length}{Specifying to which entries to apply the transformation to.}

\item{n}{The search-space dimension.}

\item{k}{The number of position-dependent parameters.}

\item{params}{List of parameters to this shape}
}
\value{
The vector x to which the shape has been applied. Those entries that were changed now have their final objective value, but more shapes can follow for the later indices.
}
\description{
wfgTransformation - used by wfgEval to apply a transformation to part of the search-space vector
}

