% Generated by roxygen2 (4.0.1): do not edit by hand
\name{withDominated}
\alias{baGraphicsWithDominated}
\alias{withDominated}
\title{withDominated}
\usage{
withDominated(func)

baGraphicsWithDominated(name, spec)
}
\arguments{
\item{func}{The function to plot}

\item{spec}{Specification of the function to plot}
}
\value{
nothing
}
\description{
withDominated is a visualization used alongside the wfg package. It shows the usual representation of objective space and pareto-frontier but complements it with the dominated individuals.\cr
baGraphicsWithDominated is a wrapper taking a filename for output and a specification for the test-function.
}
\examples{
withDominated(wfgWrap(2, c(sConvex)))
baGraphicsWithDominated("withDominatedTFlat", c(tFlat, from=0.4, to=0.9, sLinear))
}

